+++ odoo/local-src/specific_purchase/__manifest__.py
@@ -12,6 +12,7 @@
         'purchase',
         'stock',
         'calendar',
+        'purchase_cancel_reason',
     ],
     'website': 'http://www.camptocamp.com',
     'data': [
@@ -22,6 +23,7 @@
         "views/bank_holiday.xml",
         "views/purchase_config_settings.xml",
         "views/product_state.xml",
+        "views/email_template.xml",
 
         # Data
         "data/product_state.xml",